requirements.txt
celery==5.3.0
redis==4.6.0
flower==1.2.0
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
pandas==2.1.4
openpyxl==3.1.0
pydantic==2.5.0
psycopg2-binary==2.9.9